backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.2
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
redis==5.0.4
celery==5.4.0
httpx==0.27.0
openai==1.30.5
google-generativeai==0.5.4
python-multipart==0.0.9
sse-starlette==2.1.0
stripe>=8.0.0
tiktoken==0.7.0
